Feat: 스톱워치 타이머 클래스로 분리
