+++ app/src/main/java/com/android04/godfisherman/ui/main/MainViewModel.kt
@@ -11,6 +11,7 @@ import com.android04.godfisherman.data.repository.LocationRepository
 import com.android04.godfisherman.data.repository.StopwatchRepository
 import com.android04.godfisherman.localdatabase.entity.TmpFishingRecord
 import com.android04.godfisherman.utils.LocationHelper
+import com.android04.godfisherman.utils.StopwatchManager
 import com.android04.godfisherman.utils.toTimeMilliSecond
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
@@ -29,7 +30,7 @@ class MainViewModel @Inject constructor(
         var isTimeLine = false
         var isFromService = false
     }
-
+    val stopwatch = StopwatchManager({ time -> _displayTime.postValue(time.toTimeMilliSecond()) })
     val stopwatchOnFlag: MutableLiveData<Boolean> = MutableLiveData(false)
     var beforeMenuItemId: Int = 0
     var isFromStopwatchFragment: Boolean = false
@@ -59,10 +60,6 @@ class MainViewModel @Inject constructor(
     private val _tmpFishingList: MutableLiveData<List<TmpFishingRecord>> by lazy { MutableLiveData<List<TmpFishingRecord>>() }
     val tmpFishingList: LiveData<List<TmpFishingRecord>> = _tmpFishingList
 
-    private lateinit var stopwatch: Timer
-    var time = 0.0
-    var resumeTime = 0.0
-
     private val _displayTime: MutableLiveData<String> by lazy { MutableLiveData<String>("00:00:00.00") }
     val displayTime: LiveData<String> = _displayTime
 
@@ -77,43 +74,32 @@ class MainViewModel @Inject constructor(
     }
 
     fun passedTimeFromService(passedTime: Double) {
-        time = passedTime
-        stopwatch = Timer()
-        stopwatch.scheduleAtFixedRate(StopwatchTask(), 0, 10)
+        stopwatch.start(10, passedTime)
         _isStopwatchStarted.value = true
-    }
-
-    fun resetStopwatch() {
-        time = 0.0
-        _displayTime.postValue(time.toTimeMilliSecond())
     }
 
     private fun startStopwatch() {
         isTimeLine = true
-        stopwatch = Timer()
-        stopwatch.scheduleAtFixedRate(StopwatchTask(), 0, 10)
+        stopwatch.start(10)
         _isStopwatchStarted.value = true
     }
 
     fun endStopwatch() {
         isTimeLine = false
-        stopwatch.cancel()
-        resumeTime = time
+        stopwatch.end()
         _isStopwatchStarted.value = false
     }
 
     fun resumeStopwatch() {
         isTimeLine = true
-        stopwatch = Timer()
-        time = resumeTime
-        stopwatch.scheduleAtFixedRate(StopwatchTask(), 0, 10)
+        stopwatch.resumeStopwatch(10)
         _isStopwatchStarted.value = true
     }
 
     fun saveTimeLineRecord() {
         if (!_tmpFishingList.value.isNullOrEmpty()) {
             viewModelScope.launch(Dispatchers.IO) {
-                when (repository.saveTimeLineRecord(time)) {
+                when (repository.saveTimeLineRecord(stopwatch.getTime())) {
                     is Result.Success -> {
                         //todo
                     }
@@ -124,20 +110,6 @@ class MainViewModel @Inject constructor(
             }
         }
         _isAfterUpload.value = true
-        resetStopwatch()
-    }
-
-    private inner class StopwatchTask() : TimerTask() {
-        override fun run() {
-            time++
-            _displayTime.postValue(time.toTimeMilliSecond())
-        }
-
-        override fun cancel(): Boolean {
-            resumeTime = time
-            time = 0.0
-            return super.cancel()
-        }
     }
 
     fun loadTmpTimeLineRecord() {
@@ -148,11 +120,15 @@ class MainViewModel @Inject constructor(
 
     fun passStopwatchToService() {
         _isStopwatchStarted.value = false
-        stopwatch.cancel()
+        stopwatch.end()
     }
 
     fun setIsAfterUploadFalse() {
         _isAfterUpload.value = false
+    }
+
+    fun requestLocation() {
+        locationHelper.setLocationUpdate { updateLocation() }
     }
 
     private fun updateLocation() {
@@ -163,11 +139,6 @@ class MainViewModel @Inject constructor(
                 _currentLocation.postValue(location)
             }
         }
-    }
-
-    fun requestLocation() {
-        Log.d("LocationUpdate", "requestLocation() 실행")
-        locationHelper.setLocationUpdate { updateLocation() }
     }
 
 }
